Add files via upload
This dashboard was created with PowerBI, it displays information collected from a professional survey. The data was originally contained in an Excel file, and I connected it to PowerBI to create the dashboard. The survey data includes important demographic information, such as average salary, age, country of residence, and perceived difficulty of breaking into the data industry, among other information. Additionally, the survey asked questions about respondents' happiness levels in relation to work-life balance and salary. Using PowerBI, I was able to create a visually engaging and informative dashboard that displays this data in an easily digestible way. This dashboard can be used to gain insights into the demographics of those working in the data industry and to inform decisions related to hiring and compensation strategies.